backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
supabase==2.9.1
google-generativeai==0.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
Pillow==10.4.0
python-dotenv==1.0.1
